+++ apps/web/src/app/_components/footer/version.ts
@@ -1,6 +1,6 @@
 import "server-only"
 
-import packageJson from "../../../../../package.json" with { type: "json" }
+import packageJson from "../../../../../../package.json" with { type: "json" }
 
 export const version = packageJson.version
 export const isProduction = process.env.NODE_ENV === "production"
